+++ src/onegov.user/onegov/user/forms/__init__.py
@@ -1,4 +1,11 @@
 from onegov.user.forms.login import LoginForm
 from onegov.user.forms.registration import RegistrationForm
+from onegov.user.forms.reset_password import PasswordResetForm
+from onegov.user.forms.reset_password import RequestPasswordResetForm
 
-__all__ = ['LoginForm', 'RegistrationForm']
+__all__ = [
+    'LoginForm',
+    'PasswordResetForm',
+    'RegistrationForm',
+    'RequestPasswordResetForm',
+]