+++ src/onegov.user/onegov/user/utils.py
@@ -112,3 +112,28 @@ def yubikey_otp_to_serial(otp):
         value += (bytesarray[i] & 255) << (shift & mask_value)
 
     return value
+
+
+def password_reset_url(user, request, url):
+    """ Appends the token needed by PasswordResetForm for a password reset.
+
+    :user:
+        The user (model).
+
+    :request:
+        The request.
+
+    :url:
+        The URL which points to the password reset view (which is using the
+        PasswordResetForm).
+
+    :return: An URL containg the password reset token.
+
+    """
+
+    token = request.new_url_safe_token({
+        'username': user.username,
+        'modified': user.modified.isoformat() if user.modified else ''
+    })
+
+    return '{url}?token={token}'.format(url=url.rstrip('/'), token=token)