+++ src/onegov.user/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Adds forms and helper functions for password reset.
+  [msom]
+
 0.10.1 (2017-05-02)
 ~~~~~~~~~~~~~~~~~~~